# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Local Model Works
  version: 1.0.0
  description: >
    Control plane API for a self-hosted local-AI workshop. Single-operator.
    All mutations require a session cookie plus the X-CSRF-Token header.
    Errors use the structured Error shape; resource conflicts are 409,
    incompatible plans are 422, and unavailable required agents are 503.
servers:
  - url: /api/v1
...
paths:
  /login:
    post:
      tags: [auth]
      operationId: login
      summary: Exchange username/password for a session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '204':
          description: Session cookie set
        '401':
          $ref: '#/components/responses/Unauthorized'
  /logout:
    post:
      tags: [auth]
      operationId: logout
      summary: End the current session
      responses:
        '204':
          description: Session ended
        '401':
          $ref: '#/components/responses/Unauthorized'
  /session:
    get:
      tags: [auth]
      operationId: getSession
      summary: Current session metadata including the CSRF token
      responses:
        '200':
          description: Session info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /system/info:
    get:
      tags: [system]
      operationId: systemInfo
      responses:
        '200':
          description: Server identity and fleet summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /modules:
    get:
      tags: [system]
      operationId: listModules
      summary: Enabled first-party modules (descriptors present server- and client-side)
      responses:
        '200':
          description: Module descriptors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Module'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /events:
    get:
      tags: [system]
      operationId: streamEvents
      summary: Server-sent event stream (telemetry, job state, log tails)
      description: >
        Monotonic integer event IDs. Clients resume with Last-Event-ID;
        events older than the retention window are pruned and the stream
        resends from the oldest retained ID.
      parameters:
        - name: types
          in: query
          description: Comma-separated event type filter
          schema:
            type: string
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /enrollment-tokens:
    post:
      tags: [enrollment]
      operationId: createEnrollmentToken
      summary: Create a one-use, ten-minute enrollment token
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnrollmentTokenRequest'
      responses:
        '201':
          description: Token (plaintext shown once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentToken'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [enrollment]
      operationId: listEnrollmentTokens
      responses:
        '200':
          description: Token list (hashed, no plaintext)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnrollmentToken'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /enrollment-tokens/{id}:
    delete:
      tags: [enrollment]
      operationId: deleteEnrollmentToken
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '204':
          description: Token revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /migration/scan:
    post:
      tags: [migration]
      operationId: migrationScan
      summary: Start a resumable legacy DGX-Dashboard scan job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MigrationScanRequest'
      responses:
        '202':
          description: Scan job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /migration/import:
    post:
      tags: [migration]
      operationId: migrationImport
      summary: Start a digest-gated import into an isolated staging state root
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MigrationImportRequest'
      responses:
        '202':
          description: Import job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
components:
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource conflict (occupation, duplicate, stale state)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unprocessable:
      description: Incompatible plan or invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceUnavailable:
      description: Required agent(s) unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code, e.g. node.occupied, plan.incompatible
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    Session:
      type: object
      required: [username, csrf_token, expires_at]
      properties:
        username:
          type: string
        csrf_token:
          type: string
        expires_at:
          type: string
          format: date-time
    SystemInfo:
      type: object
      required: [version, commit, time, nodes, agent_url, ca_fingerprint]
      properties:
        version:
          type: string
        commit:
          type: string
        time:
          type: string
          format: date-time
        agent_url:
          type: string
          format: uri
        ca_fingerprint:
          type: string
          pattern: '^[0-9a-f]{64}$'
        nodes:
          type: object
          properties:
            total:
              type: integer
            online:
              type: integer
            approved:
              type: integer
    Module:
      type: object
      required: [id, title, route]
      properties:
        id:
          type: string
        title:
          type: string
        route:
          type: string
        nav:
          type: object
          properties:
            label:
              type: string
            order:
              type: integer
            icon:
              type: string
        settingsSchema:
          type: object
          additionalProperties: true
        jobKinds:
          type: array
          items:
            type: string
        artifactKinds:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
    CreateEnrollmentTokenRequest:
      type: object
      properties:
        description:
          type: string
    EnrollmentToken:
      type: object
      required: [id, expires_at]
      properties:
        id:
          type: string
        token:
          type: string
          description: Plaintext token; present only on creation
        description:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        used_at:
          type: string
          format: date-time
          nullable: true
...
    MigrationScanRequest:
      type: object
      required: [legacy_root, legacy_state]
      additionalProperties: false
      properties:
        legacy_root:
          type: string
        legacy_state:
          type: string
        legacy_revision:
          type: string
          pattern: '^[0-9a-f]{40}$'
        ini_path:
          type: string
        docker:
          type: boolean
          default: false
    RunAccepted:
      type: object
      required: [run_id]
      properties:
        run_id: { type: string, format: uuid }
...
    MigrationImportRequest:
      type: object
      required: [plan_digest, legacy_root, legacy_state, confirm]
      additionalProperties: false
      properties:
        plan_digest:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        legacy_root: { type: string }
        legacy_state: { type: string }
        legacy_revision:
          type: string
          pattern: '^[0-9a-f]{40}$'
        ini_path: { type: string }
        confirm: { const: true }